+++ store/provider_test.go
@@ -133,7 +133,7 @@ func Test_SingleNodeProvideLastIndex(t *testing.T) {
 
 	lm, err := provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 
 	er := executeRequestFromStrings([]string{
@@ -150,7 +150,7 @@ func Test_SingleNodeProvideLastIndex(t *testing.T) {
 
 	newLI, err := provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 	if newLI <= lm {
 		t.Fatalf("last index should have changed")
@@ -169,7 +169,7 @@ func Test_SingleNodeProvideLastIndex(t *testing.T) {
 	}
 	newLI, err = provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 	if newLI != lm {
 		t.Fatalf("last index should not have changed")
@@ -181,7 +181,7 @@ func Test_SingleNodeProvideLastIndex(t *testing.T) {
 	}
 	newLI, err = provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 	if newLI != lm {
 		t.Fatalf("last index should not have changed")
@@ -199,7 +199,7 @@ func Test_SingleNodeProvideLastIndex(t *testing.T) {
 	}
 	newLI, err = provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 	if newLI == lm {
 		t.Fatalf("last index should changed even with constraint violation")
@@ -219,7 +219,7 @@ func Test_SingleNodeProvideLastIndex(t *testing.T) {
 	}
 	newLI, err = provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 	if newLI <= lm {
 		t.Fatalf("last index should have changed")
@@ -259,7 +259,7 @@ func Test_SingleNodeProvideLastIndex_Restart(t *testing.T) {
 
 	lm, err := provider.LastIndex()
 	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
+		t.Fatalf("failed to get last index: %s", err.Error())
 	}
 	if err := s.Close(true); err != nil {
 		t.Fatalf("failed to close store: %s", err.Error())
@@ -274,11 +274,11 @@ func Test_SingleNodeProvideLastIndex_Restart(t *testing.T) {
 		t.Fatalf("Error waiting for leader: %s", err)
 	}
 	provider = NewProvider(s, false, false)
-	newLI, err := provider.LastIndex()
-	if err != nil {
-		t.Fatalf("failed to get last modified: %s", err.Error())
-	}
-	if exp, got := lm, newLI; exp != got {
-		t.Fatalf("last index should be the same after restart\nexp: %d\ngot: %d", exp, got)
-	}
+	testPoll(t, func() bool {
+		newLI, err := provider.LastIndex()
+		if err != nil {
+			t.Fatalf("failed to get last index: %s", err.Error())
+		}
+		return lm == newLI
+	}, 100*time.Millisecond, 5*time.Second)
 }